[REF] evaluation: group compilation parameters
Group all compilation parameters in one place: the Evaluator.
